+++ server/src/completion/basic.ts
@@ -1,6 +1,7 @@
 import { CompletionItem, Position, Range, TextDocumentPositionParams } from "vscode-languageserver";
 import { documents, programs } from "../server";
 import { onCommandCompletion } from "./command";
+import { onLabelCompletion } from "./label";
 
 export function onCompletionImpl(positionParams: TextDocumentPositionParams): CompletionItem[] {
   const document = documents.get(positionParams.textDocument.uri);
@@ -12,10 +13,14 @@ export function onCompletionImpl(positionParams: TextDocumentPositionParams): Co
   const splitResult = line.split(/\s+/).length;
   const program = programs.get(positionParams.textDocument.uri);
   // 1 - label
+  if (splitResult == 1)
+    return onLabelCompletion(program);
   // 2 - command
-  // 3 - after the command
   if (splitResult == 2)
     return onCommandCompletion(program);
+  // 3 - after the command
+  if (splitResult == 3)
+    return onLabelCompletion(program, true);
   return [];
 }